[refactor]: 3. 패키지 분리 완료
