+++ src/test/java/softeer2nd/chess/BoardTest.java
@@ -1,8 +1,9 @@
-package softeer2nd;
+package softeer2nd.chess;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import softeer2nd.Unit.Pawn;
+import softeer2nd.Board;
+import softeer2nd.chess.pieces.Pawn;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
